Askademia/agent_suite/requirements.txt
uagents
google-genai
python-dotenv
pydantic